requirements.txt
affine==2.4.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
numpy==2.2.5
pillow==11.2.1
pyparsing==3.2.3
rasterio==1.4.3
